add tests

+++ include/Colony.h
@@ -5,14 +5,14 @@
 class Colony{
 	protected:
 		Sequence & seq;
-		double ** pheromones;
+		long double ** pheromones;
 		bool loneInstance;
-		vector<pair<double,deque<int>>> * newPheromones;
+		vector<pair<long double,deque<int>>> * newPheromones;
 		int findInRoulette(float val, vector<float> & roulette);
 		int findInWeightedRoulette(float val, vector<float> roulette, vector<int> * weights);	
 	public:
 		Colony(Sequence & newSeq);
-		Colony(Sequence & newSeq, double ** linkedPheromones, vector<pair<double,deque<int>>> * linkedNewPheromones);
+		Colony(Sequence & newSeq, long double ** linkedPheromones, vector<pair<long double,deque<int>>> * linkedNewPheromones);
 		~Colony();
 		virtual void ant()=0;
 		virtual void pheremoneApplyEvent();
